add flatmap extension for sqldelight query

+++ common/src/commonTest/kotlin/com/twidere/twiderex/db/dao/SqlDelightDirectMessageEventDaoImplTest.kt
@@ -28,7 +28,7 @@ class SqlDelightDirectMessageEventDaoImpl : DirectMessageEventDao {
     override fun getPagingSource(
         accountKey: MicroBlogKey,
         conversationKey: MicroBlogKey
-    ): PagingSource<Int, UiDMEvent> {
+    ): PagingSource<Long, UiDMEvent> {
         TODO("Not yet implemented")
     }
